net: Extract skb_extra
