+++ src/net/skbuff/Mybuild
@@ -16,6 +16,8 @@ module skbuff {
 	source "skb.c"
 	depends embox.arch.interrupt
 	depends embox.mem.sysmalloc_api
+	
+	source "skb_extra.c"
 
 	source "skb_queue.c"
 	depends embox.arch.interrupt
